+++ ContentPatcher/Framework/Constants/PlayerType.cs
@@ -1,0 +1,12 @@
+namespace ContentPatcher.Framework.Constants
+{
+    /// <summary>A multiplayer player type.</summary>
+    internal enum PlayerType
+    {
+        /// <summary>The player hosting the world in multiplayer.</summary>
+        HostPlayer,
+
+        /// <summary>The current player, regardless of whether they're the main or secondary player.</summary>
+        CurrentPlayer
+    }
+}